POT files corrections

+++ properties2pot.sh
@@ -8,28 +8,35 @@ IFS=$'\n'
 for f in "$@"
 do
 if [[ $f == 'Bundle.properties' ]]; then
-PWD=`pwd`
+	PWD=`pwd`
 
-path=`echo "$PWD" | sed 's,.*\/src\/\(.*\)$,\1,' | sed 's,/,-,g'`
+	path=`echo "$PWD" | sed 's,.*\/src\/\(.*\)$,\1,' | sed 's,/,-,g'`
 
-if [[ $path == org-* ]]; then
-echo $path
-fname=${path}.pot
-# generate POT file from Bundle.properties
-#msginit --input=$f --properties-input --output-file=gephi-${path}.pot
-msgcat $f --properties-input --output-file=$fname
+	#rm *.pot
 
-#add header
-head -q $fname > tmp.txt
-head -q ${ROOT}/pot-header.txt -n 18 tmp.txt > $fname
-rm tmp.txt
-fi
+	if [[ $path == org-* ]]; then
+		echo $path
+		fname=${path}.pot
+		# generate POT file from Bundle.properties
+		msgcat $f --properties-input --output-file=$fname
+
+		if [[ -f $fname ]]; then
+			#add header
+			cp $fname tmp.txt
+			cat ${ROOT}/pot-header.txt tmp.txt > $fname
+			rm tmp.txt
+
+			#check file
+			msgfmt -c $fname
+			#rm messages.mo
+		fi
+	fi
 
 fi
 if [[ -d "${f}" ]]; then
-cd "${f}"
-RecurseDirs $(ls -1 ".")
-cd ..
+	cd "${f}"
+	RecurseDirs $(ls -1 ".")
+	cd ..
 fi
 done
 IFS=$oldIFS
